Start implementing a cache for album data.
Create a cache of album information in a database.

Subscribe to updates to the status of the server, paying
attention to the time that the music database was last
scanned.  If this was more recently than the cache then
clean it and repopulate it.

Currently implements the cleaning and initial creation
steps.