Provide a perceived speed up in the album cache list by scheduling a thread to run every two seconds that notifies observers of updates to the DB (if necessary) and use this when updating the album artwork.
Use a three step process for creating the icons -- first,
download to disk.  Then parse to get the dimensions, then
use the dimensions as part of the artwork icon scaling
process.  This saves some memory, and deals with pathological
artwork that's many hundreds or thousands of pixels in a given
dimension.