Push the responsibility for creating and regenerating the cache out of the Service and in to the AlbumCacheProvider.
To do this, support an onServerStateChanged() callback
that the service will call on bound clients with the
old and new server states.

This involved making SqueezerServerState parcelable so
that it can be passed to the callback.

Override SQLiteOpenHelper.getWritableDatabase() and move
the cache creation / maintenance logic in to there, so
that the cache is checked and possibly recreated every
time the database is opened