First cut at implementing 'clear album cache' and 'pre-load album cache' functionality.
Push the cache of which pages have been requested from the cursor
to the provider.

Provide a 'rebuildCache' method in the provider that can clears out
the current cache and recreates it.

Modify the service and CLI connector to use a RemoteCallbackList
for album callbacks.  This allows multiple callbacks to exist and
be called in sequence, instead of just one.

Use this new functionality in AlbumsListActivity to clear out the
cache and cause a cache pre-load.  The cache pre-load is done by
installing an additional callback handler, kicking off the first
request, and then requesting the next page via the callback and
the progress dialog.