Context menu replacements
Which things you can do with a squeezer server item, is shown by the context menu for the item, which is shown by a long press on the item.
Since many users may not even know that long presses are possible, chances are they might not discover, that different actions can be performed for a particular item.
This commit implements an icon, which is shown, if the item has other actions. The icon are displayed on item lists for each item, which has context menu actions.
When the icon is pressed, a popup menu is shown, where is context action can be selected.

- Refactored the action bar popup menu, so it can be used elsewhere.
- Added an optional context menu icon, for each view used in the item lists.
- Moved the context menu actions, from long press to the context popup menu.