UI polish for the per-item context menu images.
Make the images wider, and for icon layouts ensure it's the same size
as the icon.

Provide visual feedback that the image has been pressed by using a
state list drawable as the background. It's the same state list
drawable that is used when a list item is pressed.

The files are from $SDK_DIR/platforms/android-15/data/res/drawable*
to be consistent with the holo-dark theme the app is using.