Work in progress, using Fragments for lists.
First cut at implementing a SongListFragment that can be hosted in an
activity to show songs according to different criteria.

Implement AlbumListActivity to host this, showing all songs from a given
album.

Main change is to provide an OrderPages interface that both activities and
fragments can implement.  This lets the adapter receive an object that
implements this interface, and not need to care if it's an activity or a
fragment.  The adapter then uses this, instead of calling the itemView's
getActivity() method.

Known problems:

* Context menu for list items in the fragment doesn't work.
* List items have additional padding because it's falling back to a generic
  view.
* Lots of duplication of existing code that must be refactored.
* OrderPages interface probably contains too many methods.