Use the AutoValue extension for model/* objects.
Do this in conjunction with the AutoParcelable and AutoWith extensions to
reduce the amount of boilerplate code in each model object.

This makes each model object immutable, so it's easier to reason about state
across threads.

Each model object is constructed with a static factory method, fromMap(),
that can construct the object from a token map. It also drops the get/can
prefix from the accessors (as there are now no setters).

One issue with this change is that the previous version of the code used an
object's class name (e.g., "plugin.class") at runtime to generate the key
for intents when sending, and used the class's name (e.g., "Plugin.class")
to look up keys when receiving an intent. This doesn't work for AutoValue
objects because the class name of the instantiated object is "AutoValue_...".
So introduce intentExtraKey() for a class to return the string to use when
packing objects of that class in to an intent.

Add new tests for each model object. The tests are local unit tests. This
causes a problem for model objects that instantiate android.net.Uri objects.
To fix this, pull in the UnMock plugin library that makes certain classes
from the Android framework available for local unit tests.